# apps/api/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "snapdone Go API. 인증 credential은 JWT가 아닌 opaque 서버 세션 토큰.",
        "title": "snapdone API",
        "contact": {},
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/auth/exchange": {
            "post": {
                "description": "콜백이 넘긴 result code를 앱 verifier · state와 함께 한 번만 세션으로 바꾼다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "로그인 결과 교환",
                "parameters": [
                    {
                        "description": "result code · 앱 PKCE verifier · 앱 state",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.ExchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 · 틀린 proof · 만료 · 재사용 (invalid_callback)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 · OAuth 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/auth/oauth/callback": {
            "get": {
                "description": "provider가 브라우저를 돌려보내는 곳. 결과는 서버 설정의 앱 · web 복귀 URI로 302 redirect한다.\n성공이면 `code`(60초 · 일회용 result code)와 앱 `state`를, 실패면 `error`(cancelled · provider_unavailable · invalid_callback)와 앱 `state`를 query로 붙인다.\nprovider 토큰은 넘기지 않는다. transaction을 찾지 못하면 redirect하지 않고 400이다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "provider 콜백",
                "parameters": [
                    {
                        "type": "string",
                        "description": "서버가 만든 provider state",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "provider authorization code",
                        "name": "code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "provider 오류 (예: access_denied)",
                        "name": "error",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "복귀 URI로 redirect",
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "복귀 URI ?code=\u0026state= 또는 ?error=\u0026state="
                            }
                        }
                    },
                    "400": {
                        "description": "모르는 · 이미 쓴 state (invalid_callback)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 · OAuth 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/oauth/cancel": {
            "post": {
                "description": "앱이 취소한 시작을 폐기한다. 본문이 잘못됐거나 모르는 state여도 204다.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "로그인 취소",
                "parameters": [
                    {
                        "description": "시작할 때 보낸 앱 state",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.OAuthCancelRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "503": {
                        "description": "인증 · OAuth 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/oauth/start": {
            "post": {
                "description": "앱 PKCE challenge · state로 OAuth transaction을 만들고 provider 동의 화면 주소를 돌려준다.\nGoogle state · PKCE verifier · nonce는 앱 값과 별개로 서버가 새로 만든다. 같은 앱 state는 한 번만 쓸 수 있다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "oauth"
                ],
                "summary": "로그인 시작",
                "parameters": [
                    {
                        "description": "provider · 앱 challenge(S256) · 앱 state · platform",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.OAuthStartRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.AuthorizeURLResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 · 미설정 provider · state 재사용 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 · OAuth 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/onboarding": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "내 온보딩 단계와 사용 목적. mobile과 web이 같은 진행에서 이어 간다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "온보딩 진행",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.OnboardingResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "단계(intro · purpose · first-image)와 사용 목적을 저장한다. 목적은 first-image에서만 있고 빈 목록은 건너뜀이다.\nunsure는 혼자만 고를 수 있다. 같은 단계를 다시 저장하거나 한 단계 앞으로만 갈 수 있고, 온보딩을 마친 뒤에는 바꿀 수 없다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "onboarding"
                ],
                "summary": "온보딩 진행 저장",
                "parameters": [
                    {
                        "description": "저장할 진행",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpserver.OnboardingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpserver.OnboardingResponse"
                        }
                    },
                    "400": {
                        "description": "계약 밖의 진행 (invalid_onboarding)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "credential 없음 · 만료 · 취소 (session_expired)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "이미 마친 온보딩 (onboarding_complete) · 건너뛰거나 되돌아가는 단계 (onboarding_out_of_order)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "내부 오류 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "인증 비활성 (provider_unavailable)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httpserver.AuthorizeURLResponse": {
            "type": "object",
            "properties": {
                "authorizeUrl": {
                    "type": "string",
                    "example": "https://accounts.google.com/o/oauth2/v2/auth?..."
                }
            }
        },
...
        "httpserver.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "enum": [
                        "session_expired",
                        "provider_unavailable",
                        "invalid_callback",
                        "invalid_image",
                        "image_too_large",
                        "unsupported_image",
                        "job_not_found",
                        "invalid_onboarding",
                        "onboarding_complete",
                        "onboarding_out_of_order"
                    ],
                    "example": "session_expired"
                }
            }
        },
        "httpserver.ExchangeRequest": {
            "type": "object",
            "required": [
                "code",
                "state",
                "verifier"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "one-time-result-code"
                },
                "state": {
                    "type": "string",
                    "example": "base64url-app-state-43-to-128-chars"
                },
                "verifier": {
                    "type": "string",
                    "example": "base64url-app-verifier-43-to-128-chars"
                }
            }
        },
...
        "httpserver.LoginResponse": {
            "type": "object",
            "properties": {
                "credential": {
                    "type": "string",
                    "example": "opaque-session-credential"
                },
                "session": {
                    "$ref": "#/definitions/httpserver.SessionResponse"
                }
            }
        },
        "httpserver.OAuthCancelRequest": {
            "type": "object",
            "required": [
                "state"
            ],
            "properties": {
                "state": {
                    "type": "string",
                    "example": "base64url-app-state-43-to-128-chars"
                }
            }
        },
        "httpserver.OAuthStartRequest": {
            "type": "object",
            "required": [
                "challenge",
                "platform",
                "provider",
                "state"
            ],
            "properties": {
                "challenge": {
                    "type": "string",
                    "example": "base64url-S256-challenge-of-the-app-verifier"
                },
                "platform": {
                    "type": "string",
                    "enum": [
                        "mobile",
                        "web"
                    ],
                    "example": "mobile"
                },
                "provider": {
                    "type": "string",
                    "example": "google"
                },
                "state": {
                    "type": "string",
                    "example": "base64url-app-state-43-to-128-chars"
                }
            }
        },
        "httpserver.OnboardingRequest": {
            "type": "object",
            "required": [
                "step"
            ],
            "properties": {
                "purposes": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "food",
                            "shopping",
                            "travel",
                            "events",
                            "receipt",
                            "foreign-language",
                            "work",
                            "unsure"
                        ]
                    },
                    "example": [
                        "food",
                        "receipt"
                    ]
                },
                "step": {
                    "type": "string",
                    "enum": [
                        "intro",
                        "purpose",
                        "first-image"
                    ],
                    "example": "first-image"
                }
            }
        },
        "httpserver.OnboardingResponse": {
            "type": "object",
            "properties": {
                "purposes": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "food",
                            "shopping",
                            "travel",
                            "events",
                            "receipt",
                            "foreign-language",
                            "work",
                            "unsure"
                        ]
                    },
                    "example": [
                        "food",
                        "receipt"
                    ]
                },
                "step": {
                    "type": "string",
                    "enum": [
                        "intro",
                        "purpose",
                        "first-image",
                        "complete"
                    ],
                    "example": "first-image"
                }
            }
        },
...
        "httpserver.SessionResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-10-01T00:00:00Z"
                },
                "onboardingStep": {
                    "type": "string",
                    "enum": [
                        "intro",
                        "purpose",
                        "first-image",
                        "complete"
                    ],
                    "example": "intro"
                },
                "user": {
                    "$ref": "#/definitions/httpserver.UserResponse"
                }
            }
        },
        "httpserver.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "4f1c2a9e-0000-4000-8000-000000000000"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003copaque session credential\u003e\" 형식. 서버만 해석할 수 있는 세션 토큰.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}